requirements-dev.txt
# Dependências de desenvolvimento
# Instalar com: pip install -r requirements-dev.txt

# Formatação e linting de código
black==22.1.0
blue==0.9.1
flake8==4.0.1
isort==6.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
mccabe==0.6.1

# Testes e cobertura
pytest==8.4.1
pytest-cov==4.1.0
pytest-django==4.8.0
coverage==7.10.2
pluggy==1.6.0
iniconfig==2.1.0
tomli==2.2.1

# Documentação
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
ghp-import==2.1.0
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
Pygments==2.19.2
PyYAML==6.0.2
pyyaml_env_tag==1.1

# Utilitários de desenvolvimento
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
watchdog==6.0.0
chardet==5.2.0
click==8.2.1
mypy_extensions==1.1.0
mergedeep==1.3.4
six==1.17.0
python-dateutil==2.9.0.post0
